+++ src/server/game/Server/Packets/EquipmentSetPackets.h
@@ -94,7 +94,7 @@ namespace WorldPackets
             WorldPacket const* Write() override;
 
             uint64 GUID = 0; ///< Set Identifier
-            uint8 Reason = 0;
+            int32 Reason = 0;
         };
     }
 }
